pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TheAutomationGroupPriceParcer"
version = "0.0.1"
description = "A demo application that shows how LLm technology can be used in price-data-flows can be extracted"
authors = [
  {name = "Leo Hanhart", email = "lhanhart1996@gmail.com"},
]
dependencies = [
  "fastapi",
  "dash",
  "uvicorn",
  "pandas",
  "requests",
  "python-dotenv",
  "loguru",
  "kombu",
  "docker",
  "pytest",
  "pytest-asyncio",
  "pytest-mock",
  "openai==1.57.0",
]

[options]
package_dir = "src"
